backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
livekit
sqlalchemy
pytest
google-generativeai
pydantic-settings